pyproject.toml
[project]
name = "ai-banana-earlystage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.2",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "httpx>=0.28.0",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0",
    "python-dotenv>=1.2.1",
    "minio>=7.2.0",
    "python-multipart>=0.0.20",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ultralytics>=8.3.236",
    "uvicorn>=0.38.0",
]
